+++ NEWS.md
@@ -11,6 +11,7 @@
 - Harden request header processing.
 - Harden response header processing.
 - Filter Transfer-Encoding response header.
+- Redesign and harden HTTP response streaming.
 - Add DevContainer.